sub_modules/pollution_detector/requirements.txt
fastapi
uvicorn
python-multipart
pillow
huggingface_hub
langchain
langchain-community
pydantic
python-dotenv
requests
